Simplify GNUnet scripts to behave better on signals
The main change is to have ``gnunet-arm`` monitor a service instead of just
reading from standard output, which is problematic with Bash when running on
the background.  By having the long-running monitor process we can always send
a signal to the script and have it respond inmediately, even when on the
background.

I also removed some unnecessary exports and background starts (since
``gnunet-arm -s`` always forks), and used ``gnunet-arm`` to properly stop
services with a short timeout.  The services show some errors concerning
leaks, but they may be silenced by starting services with ``-q``.