Stop gnunet_channels::Service::async_setup on signal recv
As part of the effort to make the app exit cleanly when it receives a
signal from the user. ATM, only the
gnunet_channels::Service::async_setup function supports closing properly
and so for the app to exit cleanly the gnunet-channels submodule may
start initalizing (using the async_setup function) but shouldn't
succeed (and initialize other IO objects).

This is because making all gnunet-channels IO objects "closable" isn't a
trivial task and right now the future of GNUnet as part of ouinet is
uncertain.