Stored scripts and changes for building ouinet for windows for future reference.
build-ouinet-windows.sh is a sensible script that builds a cross compiler for windows, ouinet dependencies, and ouinet itself.
The various patch files are quick-and-dirty changes to the codebase that can make this happen. These are very much
not respectable changes, but rather the first working hack that got a build going; this would need to be transmuted
into proper fixes for various codebases if we want to make this a reality.