Fix signal handling
The trick where we called `signals.async_wait` second time after
we catch the first signal doesn't work. It's because the doing so
will create a new job for the running io_service and prevent its
`run` function from exiting.

To fix that -- after the first signal is caught -- we create a new
io_service, a new set of signals working on that new io_service and
`run` that service in a separate thread so that it doesn't block the
main one.