Implemented the first part of a bittorrent mainline DHT implementation.
The current version supports most of the DHT and routing operations, but not any
payload functionality; it supports neither bittorrent tracker operations, nor
BEP-44 data storage.

Bootstrapping support is rudimentary. It relies on a single well-known bootstrap
server, and trusts it fully. It also does not store DHT nodes in persistent
storage between runs. This should probably be improved.

This DHT implementation does not detect on its own which network interfaces it
should operate on, details of which are both application-configurable and
platform-independent. Example code contains a basic demo for linux.

Current functionality includes BEP 32, BEP 42, BEP 43, parts of BEP 45, and most
of BEP 5 minus tracker functionality. Still missing are BEP 44, BEP 51, and BEP
50.