Ported to Boost 1.67.0
With the nice side effect that we no longer need to separately
download Boost.Beast as it's part of Boost now.

TODO: In this Boost version (and I think 1.66 as well), Asio is moving
towards Networking TS compatibility

https://www.boost.org/doc/libs/1_67_0/doc/html/boost_asio/net_ts.html

and as such some of the API has changed. They still retained some
of it (and we still use it), but it's been deprecated.