Have the MitM function return the SSL ``GenericConnection``
Test code moved outside too.

When loading an HTTPS URL, sometimes "Mitm exception: stream truncated" is
triggered (mostly when accessing a site for the first time, not on reload),
but the test page always loads successfully.  Maybe there is some race between
closing the SSL connection and the plain one, or maybe the reason is that we
do not yet actually read the request inside the SSL connection.`;)`