Parse HTTPS and CONNECT request, send connection to adequate port
Use the GET target (which should be a full URL) instead of the ``Host:``
header (though they probably have the same value) to help decide which
protocol to use (HTTP or HTTPS) and the port to connect to.  Parse the CONNECT
target as host and port for the same purpose (defaulting to HTTPS port 443).

Code is a little messy now and needs some organization.

Also note that SSL is not yet used by the injector to connect to the origin.