Use ``X-Ouinet-Version`` request header to have injector act as such
Thus when a client uses the injector like a normal proxy, it does not send
this header, but when it uses it like an injector, it adds the header so that
the injector knows that it may inject the response into the cache instead of
just passing it through.

The header does not appear in the cached version of the request.