Update Docker image
This incorporates (MitM-based) HTTPS support and the Origin and Proxy request
mechanisms.  Please note that the Injector mechanism now requires the use of
the ``X-Ouinet-Version`` request header, otherwise the injector will behave as
a proxy.  Clients up-to-date with this release already include the header, but
older ones do not, so they will not contribute to injection.