Add first WebExtensions and Ouinet Android/Service code
The WebExtensions code (in webextensions/ directory) is javascript code
that acts as a web browser extension (currently testing on FireFox/Linux
and Firefox/Android). The Ouinet Android/Service code is in java and
shall act as a proxy for any Android app that wishes to use it, but ATM
mainly for Firefox/Android with that WebExtension enabled.