Parse temporary HTTP descriptors, build response from head and body
A utility function has been created to be used both by the client and injector
when retrieving content from the cache.  It takes care of fetching the HTTP
descriptor (with a temporary format) from IPFS, fetching the body from IPFS
and assembling an HTTP response with the head in the descriptor and the body.

The code builds but it is still untested.