Capture `injector` itself from the cache control when putting data
Enabling `shared_from_this()` for the `InjectorCacheControl` instance to
capture it was not sufficient since its `injector` member was just a pointer
to a reference, which did not ensure it was alive either.  Since only
`injector` is needed to proceed to the next callback, just capture it by
dereferencing the smart pointer.

(Fix provided by @inetic.)