Decide to inline descriptor in data base or link to it via IPFS
Inserting the zlib-compressed descriptor is attempted first, and if the data
base index still refuses it for being too big, the IPFS link to it is inserted
instead.

The main reason for this is to avoid having all db indexes depending on IPFS
for lookup as much as possible.  Inlining descriptors should also have a
positive performance impact by reducing lookups, but we have not assessed it
experimentally.

Currently only the BEP44-based data base index is considered able to support
descriptor inlining, however it has a limit of 1000 bencoded bytes.  Inlining
is not attempted for IPFS, and if so, probably an ASCII encoding like Base64
would be needed because of the way used to encode values in the B-tree cache.

Currently there is duplication between client and injector code, to be
refactored next.