Adapt CA certificate class for both CA and non-CA certificates
This is "el cheapo" way of implementing this: the base class contains the
machinery and interface for both cases and the particular one is selected via
a boolean parameter passed to the protected constructor by subclasses.  Even
the file name remains the same.

Later on we may want to properly split the machinery into base, CA and end
certificate implementations.