Add SSL context for client to connect to injector with TLS
The injector certificate can be placed in a specific PEM file, which is read
on startup and its certificates added to the list of valid verification
certificates for that context.

If no certificate is added to the injector PEM file, the client
will (hopefully) just use system-accepted certificates.

If the file is missing, an empty one is created.