Fix links to descriptors in db contents page
The previous links were broken in several ways (no URI authority, incorrect
path) and pointed to an external host which may not be available.  Instead of
that, the new links point to the internal `/api/descriptor` endpoint which
depends on internal data base access instead of an external site to retrieve
the decriptor, and serves proper `application/json` (which triggers JSON
pretty-printing in Firefox).