Document how to handle Docker not supporting host networking
It is only supported under Linux hosts, as indicated
[here](https://docs.docker.com/network/host/).

A special loopback address 127.7.2.1 is used under Docker instead of the more
common 127.0.0.1 to reduce the chances of clashing with other daemons
listening on local ports (e.g. IPFS on 8080).  Thus, to enable port forwarding
we can easily and safely replace 127.7.2.1 with 0.0.0.0 using the shell
container.