Bring request router code closer to cache control concepts
Rename `enable_cache` to `enable_stored` to relate it to
`CacheControl::fetch_stored` and make it look like a read-only operation flag,
not write.

Rename `responder` to `fresh_channel` to relate it to
`CacheControl::fetch_fresh`.  Do not refer to "request mechanisms" since that
concept also includes cache lookup, but rather to the concept of a channel to
reach a source of new content.