Do not provide keys to BEP44 updater
Those keys are URLs, which aren't really used by the updater.  The LRU cache
just needs a differentiating key, and the provided mutable data items to be
stored already contain a value which can be used as such key, namely the salt
value (which is currently `SHA1(URL)`), at least as long as we publish entries
signed by a single BEP44 key (otherwise we should use BEP44key+salt).