Always use I2P backup key for i2pd
Get it from legacy key if missing, always replace i2pd key with the backup
one, only save i2pd key to backup if the later is missing.

This gives priority to whatever key is in the data volume over a key that
might have been generated on i2pd bootstrap, e.g. if we want to override that
key after first container boot.