Downgrade required C++ standard back to C++14
The version of CMake in both Vagrant and Docker (i.e. Debian Stretch) does not
accept the C++17 standard, so builds broke.  Using the C++14 standard instead
results in a successful build without issues under those platforms, so I am
downgrading it until a compelling reason is provided (and documented in the
commit message) to require upgrading the aforementioned platforms.