Add and use new bittorrent::collec2 function
It has the advantage that it doesn't cancel jobs which take longer to
finish than is expected. Instead it creates new jobs on timeout and
cancels all the pending jobs only when the bulk of the job is done.

It is slower than what's currently in this branch, but I believe it's
more correct and that there is a room for improvement.