Set "updated" hook after creation of index object
It leaves a possible gap between index activation and hook readiness, but this
allows to create indexes in the `build` function while retaining ownership of
the IPFS node object in the cache client object.

The previous approach was broken because the hook created in `build` referred
to the IPFS node object which was then moved into the cache client.