Save cache file path in LRU cache element.
This allows the explicit load function to open the file, seek to the beginning
of data and read it.

However the interface is weird (passing the path and open file to value
functions).  Instead, for independent loading of data it would make much more
sense to pass in the key and open file, so the value can chooose to store data
in the open file, or use the key to find the path to a different data file.