Revert local injector cache to plain response-per-file, no LRU.
With some fixes like directory creation on item write (not on read), and
atomic creation/updating of written items.

We do not want to scan the whole cache (which may be very big) to activate the
cache on boot, but if at all possible, only use on-demand access to disk
storage for that.  However, eviction still needs to be implemented, but if
files are handled properly it can be done on an independent task.