Fix missing or broken date in fresh response to be cached.
According to RFC 7231#7.1.1.2.

This ensures that `is_semi_fresh` can find a date header when loading from the
local cache, while respecting the original date if valid.  It also keeps the
date sent to the client on first retrieval consistent with that in the
descriptor and that saved in the injector's local cache.