Fix injection ID header being sent with Proxy but not with Injector.
The `X-Ouinet-Injection-ID` is added along other Ouinet internal headers being
sent to the client but not present in the cache version.

The code was originally in the generic function for fetching a URL at the
injector, which resulted in Proxy mechanism responses containing the header,
when no injection was taking place.  Also, since the Injector mechanism
response was being passed to `ouinet::util::to_cache_response`, the field was
being dropped from those when it should be present there.