Just use `fetch_http` in injector's `fetch_fresh`.
Now reading the head of a response with `empty_body` for later forwarding
works regardless of `Content-Length`, since `fetch_http` was already able to
handle the case.

However there is no mechanism to only forward the expected amount of data, and
the forwarding operation times out even if all data was transmitted, causing
`end_of_stream` when writing back response.