Make getCaCertPath a static method so it can be called before the Ouinet client is started.
This is necessary for the Ouinet service in FF, the CA root cert needs to be setup in FF
potentially before the client has started.

Move Ouinet.Config into it's own class.
Move the setup for ouinet, copying files, etc into the same step as creating the config.
Add a unit test for the Config.