Fix half-duplex forward operation.
Now it really takes care of many more bytes it can still read from the source,
according to what was previously sent.

Also initial data in the buffer was not correctly handled since there was no
way to know it was already there, so it was getting overwritten by data from
the first read.  This has been fixed by adding an `already_read` parameter.

Since the arithmetic and buffer size adjustments used to define the proper
buffers where read and write operations take place, the `Buffer` template
argument has been removed and the `buffer` argument is always a reference to
an `asio::mutable_buffer`.