Forward output buffers outside of chunk body callback.
The `yield` token is not available to the callback, which causes the coroutine
assertion `! is_running()` to fail.

The new behavior is to save the output buffer in a variable out of the
callback, flag "end of chunk" so that the parser stops the async read
operation, then send the output buffer in the HTTP forward function.

Also fix the trivial input processing function in Proxy not keeping the output
buffer alive while `http_forward` runs.