Handle trailers with an additional callback.
It was weird (and probably incorrect) to do all the I/O inside of
`http_forward` except for the last chunk and trailers.

The function still returns the response head so it can be used e.g. to see if
the origin wants to keep the connection alive.  An alternative would be to
capture the relevant information in the head processing function and just
return void.