Use a generic Ed25519 private key in the injector.
Then use it both for signing cache content and for the BEP44 index (much in
the same way the TLS certificate is shared among TCP, uTP and I2P).

The Docker wrapper script takes care of updating the configuration file to use
option `ed25519-private-key` and renaming `bep44-*-key` to `ed25519-*-key`.