Use `X-Ouinet-Sig0` instead of `Signature`.
To avoid clashing with proposals for the standarization of head
signatures (including draft-cavage-http-signatures-11 itself).

The sequence number at the end allows the receiver to replace previous
signatures from the same `keyId` (e.g. if one signature is sent with the
initial head, then another with the trailer covering more headers in it), or
to add signatures from other `keyId`s.