Download and build CMake 3.13.4 for Vagrant and Docker.
CMake is a key part of our build process and we may want to use newer versions
than those shipped by Debian from time to time, so we just pull a precise
version of CMake much in the same way we handle our build dependency on Boost.

We may as well pull binaries directly, but the rest of build instructions are
not platform-dependent, and the CMake build step is short enough and will be
run seldom enough to afford it.  In the case of Docker, the step is placed
much towards the beginning to reduce the chance of rerunning it.