TLS first, non-secure last
When the user agent requests a non-secure website (http://...) the
client will first try to access the TLS (https://...) to the origin,
then it will try the other options like injector, injector/proxy and
d-cache. If all those fail, the client will finally try the non-secure
request to the origin.