Send block once verification would have taken place.
The block should not be sent until the next chunk with a signature for it in
its chunk extensions would arrive.

The solution is not very memory-efficient since it keeps another block-sized
buffer to save the output before the current chunk is added to the quantized
buffer, which happens to alter the data of the returned block.