Increment request_version_hdr_current 1 -> 2
This is a quick fix to account for the fact that clients currently don't
check the response_version_hdr and thus will not notify the user that it
should update the software.

Incrementing request_version_hdr_current will make injectors respond
with X-Ouinet-Errror header field to old clients (those with
request_version_hdr_current == 1) and that will trigger the update
notification.