Mark some client-to-agent responses with Ouinet protocol version.
To make it evident that they are coming from the client and not from the
origin.  Keep in mind that every response coming via Origin or Proxy is
stripped from `X-Ouinet-*` headers.

This allows a Ouinet-aware agent to take special actions on the response (like
showing specially formatted error pages or warnings).