Separate client's "Bad Request" from "Bad Gateway" responses.
The former is used for non-Ouinet responses to WebSocket and non-proxy
requests, while the later is used for correct requests which could not be
completed after trying all mechanisms (the original error message, "Not
cached", was mileading since non-cache mechanisms may have been attempted).
Only the later has `X-Ouinet-Version` and carries an `X-Ouinet-Error` with a
new value that can be parsed by the agent, so that later on additional
info (like which mechanisms were attempted) can be included.

Also the error handling functions no longer swallow the error, this is done
explicitly on calling.