Boost 1.71.0: src/util/chunk_last_x.h
Removed constructors that took templated ChunkExtensions as an argument
because it was being used for things like std::string instead of using
the constructor taking string_view. Apart from being inefficient, it
also used the Beast's detail::chunk_extensions_impl which has changed in
the recent Boost.Beast and was causing compilation failures.