Use `ResponseReader`/`ResponseWriter` for BEP5/HTTP storage.
The cache's store interface is not very clean though, since both the response
head and reader need to be passed in.  This may be avoided by having
`ResponseReader` cache the head in some way as `Session` does.

Also, `ouihash` generation on storage is missing.  The actual read/write loop
should be implemented properly in `ouinet::cache` with hash generation.

There is also the issue of moving the atomic file's lowest layer to the
`ResponseWriter`.  I checked the code and it should work (since commiting does
not operate on file contents beyond closing it), but it is not very clean.

Finally, `operator==` for `Head` and `Trailer` needs to be defined in every
module which uses `ResponseReader` or `ResponseWriter`, which is quite
cumbersome.