Pass `GenericStream` to cache store operation instead of `Session`.
Since a generic `Session` which may have already processed the head without
giving the cache a say.  The alternative would be to get the proper session
from the cache, but it looks redundant.

The session is only needed to decide whether to cache the response or not, so
decide that at the fork level to leave the whole data in the cache's tine at
its discretion.  This also allows skipping creation of the store task if the
response is not going to be cached.