Start `SigningReader`, use in unit tests.
Since it is a dummy implementation that just passes parts as they come (as a
normal `Reader` does), unit tests fail.

One difference with `session_flush_signed` is that `SigningReader` gets
injection-specific parameters by value.  Since the reader will probably be
moved into a `Session`, it may be that the reader would outlive them.