Process chunk header and body for verification.
Data is acumulated until a whole block is received, then it is verified.  If
correct, a chunk header is returned to the caller and the data is queued for
the next read part operation.

Since not all low level read part calls will result in actually returning a
part to the caller, we loop until we have something to return (or we get an
empty part).