Fix cache error code in cache control when fresher from injector.
This happens when an expired response is successfully retrieved from the cache
faster than the injector, with an etag in the response but not in the
request (e.g. because it was no longer in the browser's cache), and fetching a
newer version from the injector succeeds.

In this case none of `fresh_ec` and `cache_ec` would contain an error, which
could be considered wrong, so we force an error for cache retrieval.