Default client proxy endpoint to `127.0.0.1:8077`.
Do not override it in the example configuration file.

This was already the default in Docker-based clients.

The reason for `127.0.0.1` and not something less prone to clashes like
`127.7.2.1` (as we do in the injector) is that the client is expected to run
on many platforms (while we can assume GNU/Linux for the injector), and not
all of them support binding to addresses in `127.0.0.0/8` other than
`127.0.0.1` (e.g. macOS and other BSD-based systems).

Thus it becomes more reasonable to choose a less crowded port (8077 in this
case).