Update description and rationale of chained hashes in arch doc.
Yes, chained hashes lose the property of being able to be used for the
comparison of "prefix blocks" between different injections (as the injection
id is included in the hashing).  That responsibility has moved to the
multi-peer download protocol instead, by downloading all data hashes of blocks
beforehand and comparing them literally.