Set `ANDROID_NDK_HOME` in Android build script.
Otherwise a build which starts with no Android SDK/NDK downloads may fail with
"NDK is missing a "platforms" directory.".

The export is done in `maybe_install_sdk()` (bootstrap mode) much in the same
vein as `ANDROID_HOME` is set in `setup_deps()`, although it may be better to
set them between bootstrap and build.