Move client port opening to beginning of client start.
This reduces the chances that an agent run along the client fails to connect
to it because it is still starting.  With this change, the agent can open the
connection, but request handling will get delayed until client startup is
complete.

This also allows detecting failure to open a busy port and aborting early.