Always enable keep-alive in Proxy/Injector requests to the origin.
It is pointless to lose the chance to reuse a connection from the injector to
the origin (if the later allows it) just because the client does not want
keep-alive.  Thus we completely decouple client-injector from injector-origin
keep-alives, as they are two different connections (and Proxy  `CONNECT` is
handled separately).