Release v0.15.0.
This release fixes some Android-specific breakage when loading native
libraries, which caused issues like the client getting stuck in "starting"
state with some 64-bit devices, and getting "End of file" errors when serving
the signatures of a cached resource to another client.

For Android applications using the library, they will need to add a dependency
on ReLinker.  Please see the readme for more information.

Bugfixes:

- Use ReLinker to avoid duplicate loading of native libraries under Android
  which caused (among others) failed comparisons of error categories in HTTP
  store code.