Distinguish whether UPnP is available or not.
This slightly changes the meaning of values in the `is_upnp_active` status API
entry.  If UPnP is not available in the network, `disabled` is also
reported (before it only showed if the setup was IPv6-only), so `inactive` now
means that UPnP is available but the client was unable to add a mapping (which
may indicate some issue with UPnP setup).