Correct handling of session reader not reading anything.
This may happen when the remote end closes the connection before sending a
head, or internally when an error is caught from an Injector connection before
writing anything to UA/storage queues.  Thus do not fail with an assertion on
debug builds, and use a better error code (end of HTTP stream).