Release v0.18.0.
This release allows the Ouinet client to access its local cache even when it
has no connectivity at all, and to exchange cached content with other clients
in the same LAN when it has no upstream connection to the Internet.  This
should allow for instance to browse content previously visited (and cached) in
the same device, or to share content with close devices in cases of extreme
lack of connectivity.

Also, all operations forwarding traffic for other nodes (both at the injector
and the client) report forwarded byte counts in the log, to enable computing
some bandwidth usage statistics.

Features:

- Log counts of forwarded bytes in injector signing and proxy (plain &
  tunnel) (always), and in client content serving and bridge tunnel (with
  debug on).

Enhancements:

- Allow the operation of the cache client before bootstrapping the BitTorrent
  DHT (it can be used once bootstrapped).

- Update and simplify Guix build environment instructions.

Bug fixes:

- Fix unhandled exception in local discovery start with no connectivity.

- Other minor cancellation fixes on waiting for the start of components, in
  the full-duplex forward operation and elsewhere.