Report error when getting body size of missing local cache entry.
Instead of just returning 0.

An effect of this bug was to always consider entries in the static cache, if
missing from the internal cache, as incomplete, as their body size was
compared against 0.  This would trigger using multi-peer download, which may
delay the download or make it fail, when the good data could already have been
retrieved from the static cache.