Drop all support for flagging changed options.
Boost's option parsing machinery, because of always notifying boolean
values of the same name (even if subsequent descriptions disable notifications
for them) and notifying but not actually setting the value if it has been set
before, plus not letting know when the value is actually getting set, makes a
horrible base to track actual changes to options.

This is totally hopeless.  So just save all persistent options regardless of
their value or whether they changed and move on. `>:(`