Release v0.20.0.
The main news for this release of Ouinet is the reworking of Android library
build machinery to finally support creating signed releases and uploading them
to the Maven Central repository.  This, along with updated instructions in the
readme, should make integration of Ouinet into other apps much simpler.

Features:

- Update Gradle configuration and shell scripts to allow uploading a signed
  build of the Android library to Maven Central.

Enhancements:

- Make client local proxy and front-end endpoints configurable in Android.
  You should use in your app some non-default values to avoid clashing with
  other Ouinet-enabled apps running on the same device.

- Update readme with new instructions for integration in Android apps.

- Report errors in client writing response to cache and user agent.

- Tag long-running operations in client fetch fresh from injector.

- Log client authentication failures at the injector.