Release v0.21.8
This release improves the announcing mechanism of the control cache loop
and adds an option to control the number of simultaneous announcement
attempts. These changes will allow the desktop clients to drastically
increase the groups shared in the distribute cache. The default settings
will keep the Android devices announcing one group at a time to keep its
battery usage low.

Features:
- Add new `max-simultaneous-announcements` option, which defines the number
of simultaneous BEP5 announcements to the DHT. Default for Android
devices is 1 and for the desktop ones is 16.

Enhancements:
- Spawning batches of announcement as co-routines instead of processing them
sequentially.