Release v0.21.10
This minor release fixes a bug with the Ouinet notification not appearing upon
startup, reported in censorship-no/ceno-browser#53. It was found that this bug
was caused by an attempt to restart ouinet being triggered too soon after
startup due to a perceived change in network connectivity.

Improvements:

- Update to ConnectivityStateMonitor which implements the NetworkCallback class
as a replacement for the ConnectivityBroadcastReceiver.
- Use the ConnectivityStateMonitor to more accurately track when existing
connections are lost and new connections are available.
- Avoid attempting to restart ouinet on the first network availability, or
before ouinet startup is complete, or after a 10 second timeout after initial
ouinet startup request.
- Remove unused broadcastReceivers to avoid possible conflicts/confusion with
receivers and monitors implemented in new kotlin portion of code.