Use built libraries only for Ouinet to avoid breaking other packages
Using `ldconfig` when copying `libgcrypt` and `libgpg-error` to
`/usr/local/lib` was breaking `apt` in the newer Docker images that are
using Debian 12, because their versions are incompatible.

A solution for this, is to copy the libraries to another directory and
set `LD_LIBRARY_PATH` before invoking the Ouinet binaries.