Patch filesystem in Boost 1.77.0 when building it for Android
Some devices with Android 9 and 10 were crashing when the
`boost::filesystem` library was trying to check if a file existed using
the `statx` instruction, which is blacklisted by the seccomp extension
of the Linux kernel.

This patch is a back port of the latest version of Boost and should be
removed when Ouinet is upgraded to support Boost v1.79.0.