Release v0.25.1
This release fixes the issues found in some devices with Android 9 and 10
that were crashing when invoking a CPU instruction that was in the block
list of the `seccomp` Linux kernel extension. It also contains some
improvements and minor fixes to the CI jobs.

Fixed

- Crashes in Android 9 and 10 devices that were caused by `boost::filesystem`
when attempting to use `statx` instructions.
- Dependencies of the CI jobs were changed as they were still using names
related to the deprecated Debian 10 pipelines.

Deprecated

- Configured Debian 12 pipelines using Boost 1.71 and Boost 1.74 to run only
when are manually triggered.
- All the Debian 10 pipelines were also marked only for manual execution.