Replace refs to Boost's FS ifstream/ofstream with `boost::nowide::fstream`
References to deprecated classes `ifstream` and `ofstream` in Boost's
filesystem namespace were causing compilation issues when building
Ouinet with Boost 1.79.

The usage of `boost::nowide::fstream` was preferred over native
`std::ofstream/ifstream` mainly because it's a drop-in replacement which
reduces the impact of the fix and also because using `nowide` allow us to
support non-ascii paths in Windows platforms.