Add a return to prevent that yield is passed to async_read_some
Adding return in front of `this->async_read_some_at` should ensure
that the compilation fails if someone tries to pass yield into
`async_read_some`.

This is because if `handler` is a callback, then `async_read_some_at`
will return `void`, which matches the return type of `async_read_some`.

On the other hand, if `handler` is a `yield` then `async_read_some_at`
will return `size_t` which will not match with the return type of
`async_read_some`.