Release v0.30.1
This release modifies the stop() method of the DHT instances to trigger a
cancel signal, finish the related co-routines and release the shared pointers
to the DHT that were preventing it from being destroyed.

Fixed

- A `_cancel` signal is now triggered when a DHT instance is stopped, instead
of waiting until its destructor is invoked. This prevents the spawning of
coroutines when the client process is stopped and also releases pending
locks on the shared pointer to the DHT instance, reducing the time needed to
perform a clean shutdown.