Refactor iOS and macOS cmake build to be off-source
Rewrites build-macos.sh and renames it to build-darwin.sh. The script now
uses a similar style to build-android.sh to allow building for all PLATFORMs
off-source with limited hard coding of directories. Removes build-macos
directory from .gitignore, since the builds can now be run outside of the
source directory.

This also introduces the iOS framework build and the xcframework build
which creates a library that can be used on any arm64, darwin-based platform.
The output library is now simply named ouinet.framework.