Skip the target `test-persistent-lru-cache` in Windows builds
This test is failing in Windows due to permissions in the file when the
LRU mechanism is trying to read a file while another coroutine is
writing on it (that's the expected behavior and it works fine in POSIX).

As this test covers code that's not used in production it's safe to
disable its build for now.