Release v1.4.0
Added

- Two new JNI methods `getProxyEndpoint` and `getFrontendEndpoint` that are
useful to connect with the Ouinet service when the port selection is delegated
to the OS. !125

Fixed

- Portal functions of the front-end were changed to pass and check `upnps_ptr`,
instead of attempting to de-reference it on the fly, just to prevent issues
when the pointer is null. !131

Removed

- Remove obsolete code from `OuinetNotification` and its related components.
MR !129

Security

- Add authentication for the Ouinet proxy requests and also exposes methods
in JNI that can be used in Android. MR !126
- Improve the validations applied to determine if a target is local, private
or public, which determines if it's allowed to be injected. MR !127
- Set explicit limits to the decoding of bencode strings, integers, lists
and dictionaries. MR !130